requirements.txt
amqp==5.3.1
asgiref==3.10.0
async-timeout==5.0.1
billiard==4.2.2
celery==5.5.3
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
Django==5.2.7
djangorestframework==3.16.1
gunicorn==22.0.0
kombu==5.5.4
packaging==25.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
redis==6.4.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.14
django-filter==25.2

# Additional dependencies for Phase 1
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0

# Reporting and exports
reportlab==4.0.7
openpyxl==3.1.2
Pillow==10.1.0

# API Documentation
drf-spectacular==0.27.0
